Добавил RTK Query, добавил метод запроса пользователя по id, вывел данные профиля
