+++ client/models/user.ts
@@ -8,3 +8,9 @@ export interface IUserData {
   phone?: string
   provider?: string
 }
+
+export interface IUserById extends IUserData {
+  followers?: number
+  following?: number
+  about?: string
+}
