+++ client/package.json
@@ -9,19 +9,23 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@reduxjs/toolkit": "^1.8.5",
     "@types/jwt-decode": "^3.1.0",
     "@types/qs": "^6.9.6",
     "@types/styled-components": "^5.1.10",
     "axios": "^0.21.1",
+    "classnames": "^2.3.1",
     "cookies-next": "^2.1.1",
     "jwt-decode": "^3.1.2",
     "next": "^12.2.5",
+    "next-redux-wrapper": "^7.0.5",
     "qs": "^6.10.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-line-awesome": "^1.3.2",
     "react-number-format": "^4.6.3",
     "react-preloader-icon": "^1.0.0",
+    "react-redux": "^8.0.2",
     "styled-components": "^5.3.5",
     "styled-reset": "^4.3.4"
   },
