+++ client/pages/_app.tsx
@@ -1,12 +1,14 @@
 import '../styles/globals.css'
 import type { AppProps } from 'next/app'
-import Layout from 'layout'
+import Layout from 'src/layout'
+import { wrapper } from '../src/redux/store'
 
-const MyApp = ({ Component, pageProps }: AppProps) => {
+const MyApp = wrapper.withRedux(({ Component, pageProps }: AppProps) => {
   return (
     <Layout>
       <Component {...pageProps} />
     </Layout>
   )
-}
+})
+
 export default MyApp
