+++ client/pages/_document.tsx
@@ -4,11 +4,14 @@ import Document, {
   Head,
   Main,
   NextScript,
+  DocumentInitialProps,
 } from 'next/document'
 import { ServerStyleSheet } from 'styled-components'
 
 export default class MyDocument extends Document {
-  static async getInitialProps(ctx: DocumentContext) {
+  static async getInitialProps(
+    ctx: DocumentContext,
+  ): Promise<DocumentInitialProps> {
     const sheet = new ServerStyleSheet()
     const originalRenderPage = ctx.renderPage
