+++ client/pages/rooms/[id].tsx
@@ -1,8 +1,6 @@
-import RoomPage from 'components/Room'
-import { apiV1 } from '../../core/request'
+import RoomPage from 'src/components/Room'
 import { IConversationCard } from '../../models/room'
-import { GetServerSideProps, NextPage } from 'next'
-import axios from 'axios'
+import { NextPage } from 'next'
 
 interface IProps {
   room: Pick<IConversationCard, 'title' | 'guests' | 'id'>
@@ -13,20 +11,20 @@ const Room: NextPage<IProps> = ({ room }) => {
 }
 //TODO: Поправить типизацию
 //@ts-ignore
-export const getServerSideProps: GetServerSideProps = async context => {
-  try {
-    const { id } = context.query
-    const { data } = await axios.get<Array<IConversationCard>>(
-      'http://localhost:3000/rooms.json',
-    )
-    const room = data.find(room => room.id === id)
-
-    return {
-      props: {
-        room,
-      },
-    }
-  } catch (e) {}
-}
+// export const getServerSideProps: GetServerSideProps = async context => {
+//   try {
+//     const { id } = context.query
+//     const { data } = await axios.get<Array<IConversationCard>>(
+//       'http://localhost:3000/rooms.json',
+//     )
+//     const room = data.find(room => room.id === id)
+//
+//     return {
+//       props: {
+//         room,
+//       },
+//     }
+//   } catch (e) {}
+// }
 
 export default Room
