+++ client/src/components/Logo/index.tsx
