+++ client/src/components/StepsBlock/index.tsx
