+++ client/src/hooks/useTypedDispatch.ts
@@ -1,0 +1,4 @@
+import { AppDispatch } from '../redux/store'
+import { useDispatch } from 'react-redux'
+
+export const useTypedDispatch = () => useDispatch<AppDispatch>()
