+++ client/src/hooks/useTypedSelector.ts
@@ -1,0 +1,4 @@
+import { TypedUseSelectorHook, useSelector } from 'react-redux'
+import { RootState } from '../redux/store'
+
+export const useTypedSelector: TypedUseSelectorHook<RootState> = useSelector
