+++ client/src/utils/checkJwt.ts
@@ -1,0 +1,11 @@
+import jwtDecode from 'jwt-decode'
+
+export const verify = async (token: string): Promise<boolean> => {
+  const currentTimestamp = new Date().getTime() / 1000
+  if (!token) return false
+
+  // @ts-ignore
+  const { exp } = await jwtDecode(token)
+
+  return exp > currentTimestamp
+}
