+++ client/styles/theme.ts
@@ -1,0 +1,8 @@
+export const theme = {
+  colors: {
+    primaryText: '#4f4f4f',
+    secondaryText: '#4070ff',
+    primaryButtonColor: '#6881e7',
+    secondaryButtonColor: '',
+  },
+}
